fix(containerNamespace): support secret_environment_variable state
